App-Restaurant/requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
psycopg2==2.9.7
SQLAlchemy==2.0.20
python-dotenv==1.0.0
